Removed old files. Got rid of old RESTful Raspbery API. Moved everything over to Iris RESTful API Fixed bugs in Iris caused by switching out Raspberry functions for new Roots functions.
