+++ iris/components/iris/default.htm
@@ -1,6 +1,7 @@
 <div class="iris-widget">
     <script type="text/javascript">
             var graphData = {{ graphData|raw}};
+            var rawData = {{ rawData|raw}};
             var studentId = {{userId}};
             var courseId = {{courseId}};
     </script>
