--- raspberry/controllers/Test.php
@@ -1,13 +1,0 @@
-<?php namespace Delphinium\Raspberry\Controllers;
-
-use BackendMenu;
-use BackendAuth;
-use Illuminate\Routing\Controller;
-
-class Test extends Controller {
-	
-	public function index()
-	{
-		return "Hi, API";
-	}
-}
